+++ pywren_ibm_cloud/libs/openwhisk/client.py
@@ -17,47 +17,54 @@
 import ssl
 import json
 import base64
+import urllib3
 import logging
 import requests
 import http.client
 from urllib.parse import urlparse
+from urllib3.exceptions import InsecureRequestWarning
 
+
+urllib3.disable_warnings(InsecureRequestWarning)
 logger = logging.getLogger(__name__)
 
 
-class CloudFunctionsClient:
+class OpenWhiskClient:
 
-    def __init__(self, endpoint, namespace, namespace_id=None, api_key=None,
-                 iam_api_key=None, token_manager=None, user_agent=None):
+    def __init__(self, endpoint, namespace, api_key=None, auth=None, insecure=False, user_agent=None):
         """
-        CloudFunctionsClient Constructor
+        OpenWhiskClient Constructor
+
+        :param endpoint: OpenWhisk endpoint.
+        :param namespace: User namespace.
+        :param api_key: User AUTH Key.  HTTP Basic authentication.
+        :param auth: Authorization token string "Basic eyJraWQiOiIyMDE5MDcyNCIsImFsZ...".
+        :param insecure: Insecure backend. Disable cert verification.
+        :param user_agent: User agent on requests.
         """
         self.endpoint = endpoint.replace('http:', 'https:')
         self.namespace = namespace
-        self.namespace_id = namespace_id
         self.api_key = api_key
-        self.iam_api_key = iam_api_key
-        self.token_manager = token_manager
+        self.auth = auth
 
         if self.api_key:
             api_key = str.encode(self.api_key)
             auth_token = base64.encodebytes(api_key).replace(b'\n', b'')
-            auth = 'Basic %s' % auth_token.decode('UTF-8')
-            self.effective_namespace = self.namespace
-
-        elif self.token_manager:
-            auth_token = token_manager._token
-            auth = 'Bearer ' + auth_token
-            self.effective_namespace = self.namespace_id
+            self.auth = 'Basic %s' % auth_token.decode('UTF-8')
 
         self.session = requests.session()
-        default_user_agent = self.session.headers['User-Agent']
+
+        if insecure:
+            self.session.verify = False
 
         self.headers = {
             'content-type': 'application/json',
-            'Authorization': auth,
-            'User-Agent': default_user_agent + ' {}'.format(user_agent)
+            'Authorization': self.auth,
         }
+
+        if user_agent:
+            default_user_agent = self.session.headers['User-Agent']
+            self.headers['User-Agent'] = default_user_agent + ' {}'.format(user_agent)
 
         self.session.headers.update(self.headers)
         adapter = requests.adapters.HTTPAdapter()
@@ -83,7 +90,7 @@ class CloudFunctionsClient:
         data['exec'] = cfexec
 
         logger.debug('I am about to create a new cloud function action: {}'.format(action_name))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package,
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package,
                         action_name + "?overwrite=" + str(overwrite)])
 
         res = self.session.put(url, json=data)
@@ -100,7 +107,7 @@ class CloudFunctionsClient:
         Get an IBM Cloud Functions action
         """
         logger.debug("I am about to get a cloud function action: {}".format(action_name))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package, action_name])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package, action_name])
         res = self.session.get(url)
         return res.json()
 
@@ -109,7 +116,7 @@ class CloudFunctionsClient:
         List all IBM Cloud Functions actions in a package
         """
         logger.debug("I am about to list all actions from: {}".format(package))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package, ''])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package, ''])
         res = self.session.get(url)
         if res.status_code == 200:
             return res.json()
@@ -121,7 +128,7 @@ class CloudFunctionsClient:
         Delete an IBM Cloud Function
         """
         logger.debug("Delete cloud function action: {}".format(action_name))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package, action_name])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package, action_name])
         res = self.session.delete(url)
         resp_text = res.json()
 
@@ -130,7 +137,7 @@ class CloudFunctionsClient:
 
     def update_memory(self, package, action_name, memory):
         logger.debug('I am about to update the memory of the {} action to {}'.format(action_name, memory))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace,
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace,
                         'actions', package, action_name + "?overwrite=True"])
 
         data = {"limits": {"memory": memory}}
@@ -147,7 +154,7 @@ class CloudFunctionsClient:
         List all IBM Cloud Functions packages
         """
         logger.debug('I am about to list all the IBM CF packages')
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'packages'])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'packages'])
 
         res = self.session.get(url)
 
@@ -162,7 +169,7 @@ class CloudFunctionsClient:
         Delete an IBM Cloud Functions package
         """
         logger.debug("I am about to delete the package: {}".format(package))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'packages', package])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'packages', package])
         res = self.session.delete(url)
         resp_text = res.json()
 
@@ -176,7 +183,7 @@ class CloudFunctionsClient:
         Create an IBM Cloud Functions package
         """
         logger.debug('I am about to create the package {}'.format(package))
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'packages', package + "?overwrite=False"])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'packages', package + "?overwrite=False"])
 
         data = {"name": package}
         res = self.session.put(url, json=data)
@@ -187,16 +194,16 @@ class CloudFunctionsClient:
         else:
             logger.debug("OK --> Created package {}".format(package))
 
-    def invoke(self, package, action_name, payload={}, is_remote_cluster=False, self_invoked=False):
+    def invoke(self, package, action_name, payload={}, is_ow_action=False, self_invoked=False):
         """
         Invoke an IBM Cloud Function by using new request.
         """
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions', package, action_name])
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions', package, action_name])
         parsed_url = urlparse(url)
 
         try:
-            if is_remote_cluster:
-                resp = self.session.post(url, json=payload)
+            if is_ow_action:
+                resp = self.session.post(url, json=payload, verify=False)
                 resp_status = resp.status_code
                 data = resp.json()
             else:
@@ -210,11 +217,11 @@ class CloudFunctionsClient:
                 data = json.loads(resp.read().decode("utf-8"))
                 conn.close()
         except Exception as e:
-            if not is_remote_cluster:
+            if not is_ow_action:
                 conn.close()
             if self_invoked:
                 return None
-            return self.invoke(package, action_name, payload, is_remote_cluster=is_remote_cluster, self_invoked=True)
+            return self.invoke(package, action_name, payload, is_ow_action=is_ow_action, self_invoked=True)
 
         if resp_status == 202 and 'activationId' in data:
             return data["activationId"]
@@ -233,7 +240,7 @@ class CloudFunctionsClient:
         """
         Invoke an IBM Cloud Function waiting for the result.
         """
-        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.effective_namespace, 'actions',
+        url = '/'.join([self.endpoint, 'api', 'v1', 'namespaces', self.namespace, 'actions',
                         package, action_name + "?blocking=true&result=true"])
         resp = self.session.post(url, json=payload)
         result = resp.json()